Fix users not being allowed to have different display names in different rooms.
Make User.rooms a to-one relationship and rename to room. Add a second constraint on User for their room property. Add a cascade rule on Room.members. Add room parameter when getting a user from DataController, and include this where necessary elsewhere. Add UserTests with a test for the new constraint.